python/requirements.txt
docker==3.5.1
PyYAML==3.13
